macOS: ConnectionVC: Fix support contacts
Previously, support contacts were shown in a NSTextField, which
did not handle links correctly. This commit changes that to an
NSTextView. Since there doesn't seem to be a way to add a
non-scrolling NSTextView from within Interface Builder, the
NSTextView is configured and added to the stack view in code.