iOS: Avoid updating the main table view when it's not on the top
When the search vc fetches discovery data, it is used to refresh what's
displayed in the main vc. If the main vc's table view gets updated while
the search vc is "on top of" the main vc (in the navigation stack), then
it results in a warning on the console saying: "UITableView was told to
layout its visible cells and other contents without being in the view
hierarchy".

So when the main vc is not visible, if a change is to be made, it's
postponed until the main vc becomes visible again.