Move showing of session expiry alert to ConnectionVC
When the device wakes up from sleep, we show the session-about-to-expire
alert if that was meant to be shown while the device was asleep.

This functionality was previously coded in the AppDelegate and the
NotificationCenter. In this commit, this is consolidated in the
ConnectionVC, so that we can keep track of all session-expiry alerts in
one place, and not attempt to show an alert on this if one is already
being shown.