Use the OpenVPNAdapter interface
The API for getting the log and connection info has changed in TunnelKit 5.x
and doesn't use app messages to query the tunnel process. We'd like to continue
using app messages to query the tunnel process, so we've created an adapter API
for TunnelKit (similar to WireGuardKit's adapter API) in our fork of TunnelKit.

Viewing log and connection info doesn't work yet.