For Dev ID Distribution, show a help view instead of the log
We don't have a shared location between the app and the system
extension to write the log, so we instead rely on the OS-provided
unified logging.

So when the user clicks on "View Log", we show a view that shows
how to access the log using Console.app or log show.