macOS: Notifications: Show an info alert about macOS notification
If the user chooses "Notify" when we show the alert asking if
the user wants to be notified, then show an alert telling the user
how to "Allow" notifications in macOS (by clicking on "Allow" in
the notification from macOS).

If the user interacts with that notification, the alert is dismissed
automatically.